filter-repo: use new --date-format=raw-permissive fast-import option
fast-import gained a new raw-permissive date format explictly for
allowing people to import repositories as-is.  Make use of the flag, and
stop rewriting the bogus timezone found in rails.git.

If users do not like these bogus times, they can of course write a
filter to fix them (or even make them bogus in a different way).  For
example:

    git filter-repo ... --commit-callback '
      if commit.author_date.endswith(b"+051800"):
        commit.author_date.replace(b"+051800", b"+0261")
    '

FIXME: Need to figure out what stable thing I can search for in git
output to see if it actually has the flag.  `git fast-import -h` does
NOT document it, and `git fast-import --help` only shows manpages on
some platforms...

Signed-off-by: Elijah Newren <newren@gmail.com>